added packages in the model and restructured what classes are in what package
