+++ src/main/java/com/zcrabblers/zcrabble/Controller/BoardController.java
@@ -1,6 +1,8 @@
 package com.zcrabblers.zcrabble.Controller;
 
 import com.zcrabblers.zcrabble.Model.*;
+import com.zcrabblers.zcrabble.Model.Game.Game;
+import com.zcrabblers.zcrabble.Model.Observers.ILetterObservable;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.fxml.FXML;
