+++ src/main/java/com/zcrabblers/zcrabble/Model/Observers/ITurnObservable.java
@@ -1,4 +1,4 @@
-package com.zcrabblers.zcrabble.Model;
+package com.zcrabblers.zcrabble.Model.Observers;
 
 public interface ITurnObservable {
     boolean endTurn();
