+++ src/main/java/com/zcrabblers/zcrabble/Tests/CellTests.java
@@ -1,7 +1,7 @@
 package com.zcrabblers.zcrabble.Tests;
 
-import com.zcrabblers.zcrabble.Model.Cell;
-import com.zcrabblers.zcrabble.Model.Tile;
+import com.zcrabblers.zcrabble.Model.Game.Cell;
+import com.zcrabblers.zcrabble.Model.Game.Tile;
 import org.junit.Test;
 import static org.junit.Assert.*;
