+++ src/main/java/com/zcrabblers/zcrabble/Tests/CellTupleTests.java
@@ -1,8 +1,8 @@
 package com.zcrabblers.zcrabble.Tests;
 
-import com.zcrabblers.zcrabble.Model.Cell;
+import com.zcrabblers.zcrabble.Model.Game.Cell;
 import com.zcrabblers.zcrabble.Model.CellTuple;
-import com.zcrabblers.zcrabble.Model.Tile;
+import com.zcrabblers.zcrabble.Model.Game.Tile;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
