+++ src/main/java/com/zcrabblers/zcrabble/Tests/RackTests.java
@@ -1,8 +1,8 @@
 package com.zcrabblers.zcrabble.Tests;
 
-import com.zcrabblers.zcrabble.Model.Rack;
-import com.zcrabblers.zcrabble.Model.Tile;
-import com.zcrabblers.zcrabble.Model.TileBag;
+import com.zcrabblers.zcrabble.Model.Game.Rack;
+import com.zcrabblers.zcrabble.Model.Game.Tile;
+import com.zcrabblers.zcrabble.Model.Game.TileBag;
 import org.junit.Test;
 import static org.junit.Assert.*;
