+++ src/main/java/com/zcrabblers/zcrabble/Tests/TileBagTests.java
@@ -1,5 +1,5 @@
 package com.zcrabblers.zcrabble.Tests;
-import com.zcrabblers.zcrabble.Model.TileBag;
+import com.zcrabblers.zcrabble.Model.Game.TileBag;
 import org.junit.Test;
 import java.io.FileNotFoundException;
