+++ src/main/java/module-info.java
@@ -12,4 +12,7 @@ module com.zcrabblers.zcrabble {
     exports com.zcrabblers.zcrabble.Controller;
     exports com.zcrabblers.zcrabble.Model;
     exports com.zcrabblers.zcrabble.Utils;
+    exports com.zcrabblers.zcrabble.Model.Players;
+    exports com.zcrabblers.zcrabble.Model.Observers;
+    exports com.zcrabblers.zcrabble.Model.Game;
 }
